# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Production Pulse API — film production management for filmmakers and students
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects/{projectId}/shoot-days:
    get:
      operationId: listShootDays
      tags: [shootDays]
      summary: List shoot days for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of shoot days with their scheduled scenes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ShootDay"
    post:
      operationId: createShootDay
      tags: [shootDays]
      summary: Create a new shoot day
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShootDayInput"
      responses:
        "201":
          description: Created shoot day
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShootDay"
  /projects/{projectId}/shoot-days/{id}:
    patch:
      operationId: updateShootDay
      tags: [shootDays]
      summary: Update a shoot day (rename, change date)
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShootDayPatch"
      responses:
        "200":
          description: Updated shoot day
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShootDay"
    delete:
      operationId: deleteShootDay
      tags: [shootDays]
      summary: Delete a shoot day (unschedules its scenes)
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /projects/{projectId}/shoot-days/{id}/schedule:
    put:
      operationId: setDaySchedule
      tags: [shootDays]
      summary: Set the full list of scenes for a shoot day (replaces existing)
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SceneScheduleInput"
      responses:
        "200":
          description: Updated shoot day with scheduled scenes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShootDay"
  /projects/{projectId}/shoot-days/{id}/call-sheet:
    get:
      operationId: getCallSheet
      tags: [shootDays]
      summary: Generate call sheet for a shoot day
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Generated call sheet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CallSheet"
  /projects/{projectId}/shoot-days/{id}/ad-review:
    post:
      operationId: requestAdReview
      tags: [shootDays]
      summary: Request AI First AD review of a call sheet
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: AD review feedback
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdReview"
        "503":
          description: AI unavailable
  /projects/{projectId}/schedule-warnings:
    get:
      operationId: getScheduleWarnings
      tags: [shootDays]
      summary: Get all scheduling conflict warnings for the project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of schedule warnings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScheduleWarning"

  # ── Takes (Production stage) ──────────────────────────────────────────────
...
components:
  schemas:
...
    Scene:
      type: object
      required: [id, projectId, sceneNumber, heading, intExt, dayNight, location, summary, characters, props, costumes, aiGenerated, corrected, createdAt, updatedAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        sceneNumber:
          type: integer
        heading:
          type: string
        intExt:
          type: string
          enum: [INT, EXT, INT/EXT]
        dayNight:
          type: string
          enum: [DAY, NIGHT, DAWN, DUSK, CONTINUOUS]
        location:
          type: string
        durationPages:
          type: ["number", "null"]
        summary:
          type: string
        characters:
          type: array
          items:
            type: string
        props:
          type: array
          items:
            type: string
        costumes:
          type: array
          items:
            type: string
        notes:
          type: ["string", "null"]
        aiGenerated:
          type: boolean
        corrected:
          type: boolean
        shootDayId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ShootDay:
      type: object
      required: [id, projectId, dayNumber, label, scheduledSceneIds, createdAt, updatedAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        dayNumber:
          type: integer
        label:
          type: string
        date:
          type: ["string", "null"]
          format: date
        scheduledSceneIds:
          type: array
          items:
            type: integer
        scenes:
          type: array
          items:
            $ref: "#/components/schemas/Scene"
        warnings:
          type: array
          items:
            $ref: "#/components/schemas/ScheduleWarning"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ShootDayInput:
      type: object
      required: [label]
      properties:
        label:
          type: string
          minLength: 1
        date:
          type: string
          format: date
    ShootDayPatch:
      type: object
      properties:
        label:
          type: string
        date:
          type: string
          format: date
    SceneScheduleInput:
      type: object
      required: [sceneIds]
      properties:
        sceneIds:
          type: array
          items:
            type: integer
    ScheduleWarning:
      type: object
      required: [type, severity, message, shootDayId]
      properties:
        type:
          type: string
          enum: [idle_actor, excessive_locations, page_count_overrun]
        severity:
          type: string
          enum: [warning, error]
        message:
          type: string
        shootDayId:
          type: integer
        affectedActors:
          type: array
          items:
            type: string
        affectedLocations:
          type: array
          items:
            type: string
    CallSheet:
      type: object
      required: [shootDayId, dayNumber, label, scenes, castCalls, crewCallTime]
      properties:
        shootDayId:
          type: integer
        dayNumber:
          type: integer
        label:
          type: string
        date:
          type: ["string", "null"]
        scenes:
          type: array
          items:
            $ref: "#/components/schemas/Scene"
        castCalls:
          type: array
          items:
            $ref: "#/components/schemas/CastCall"
        crewCallTime:
          type: string
        generalCallTime:
          type: string
    CastCall:
      type: object
      required: [character, callTime, onSetTime]
      properties:
        character:
          type: string
        callTime:
          type: string
        makeupTime:
          type: ["string", "null"]
        onSetTime:
          type: string
    AdReview:
      type: object
      required: [verdict, issues, coachingNote, aiAvailable]
      properties:
        verdict:
          type: string
        issues:
          type: array
          items:
            $ref: "#/components/schemas/AdIssue"
        coachingNote:
          type: string
        aiAvailable:
          type: boolean
    AdIssue:
      type: object
      required: [description, severity]
      properties:
        description:
          type: string
        severity:
          type: string
          enum: [info, warning, error]

    # ── Takes ────────────────────────────────────────────────────────────────
...